Trying to OOPify BangScores for future development #2

+++ src/main/java/DiscordBot/commands/BangScores.java
@@ -12,8 +12,8 @@ public class BangScores {
         BangHighScores highScores = GetBangScores.getBangScores(channel, path3, guild);
 
         if (highScores != null){
-            channel.sendMessage("The player with the most attempts is "+highScores.mostAttemptsPlayer+" with a score of "+highScores.attemptCount).queue();
-            channel.sendMessage("The player with the most deaths is "+highScores.mostDeathsPlayer+" with a score of "+highScores.deathCount).queue();
+            channel.sendMessage("The player with the most attempts is "+highScores.mostAttemptsPlayer+" with a score of "+(int)highScores.attemptCount).queue();
+            channel.sendMessage("The player with the most deaths is "+highScores.mostDeathsPlayer+" with a score of "+(int)highScores.deathCount).queue();
             channel.sendMessage("The player with the best survival rate is "+highScores.luckiest+" at "+highscores.bestRate+"%").queue();
             channel.sendMessage("The player with the worst survival rate is "+highScores.unluckiest+" at "+highscores.worstRate+"%").queue();
         }
